+++ dbms/src/AggregateFunctions/CMakeLists.txt
@@ -21,8 +21,8 @@ list(REMOVE_ITEM clickhouse_aggregate_functions_headers
 
 add_library(clickhouse_aggregate_functions ${clickhouse_aggregate_functions_sources})
 target_link_libraries(clickhouse_aggregate_functions PRIVATE dbms PUBLIC ${CITYHASH_LIBRARIES})
-target_include_directories (clickhouse_aggregate_functions BEFORE PRIVATE ${COMMON_INCLUDE_DIR})
+target_include_directories(clickhouse_aggregate_functions PRIVATE ${COMMON_INCLUDE_DIR})
 
-if (ENABLE_TESTS)
-    add_subdirectory (tests)
-endif ()
+if(ENABLE_TESTS)
+    add_subdirectory(tests)
+endif()
