+++ dbms/src/Functions/URL/URLHierarchy.cpp
@@ -1,19 +1,19 @@
 #include <Functions/FunctionFactory.h>
 #include <Functions/FunctionsStringArray.h>
-#include <Functions/FunctionsURL.h>
 
 namespace DB
 {
 
-class URLHierarchyImpl
+class URLPathHierarchyImpl
 {
 private:
     Pos begin;
     Pos pos;
     Pos end;
+    Pos start;
 
 public:
-    static constexpr auto name = "URLHierarchy";
+    static constexpr auto name = "URLPathHierarchy";
     static String getName() { return name; }
 
     static size_t getNumberOfArguments() { return 1; }
@@ -37,6 +37,7 @@ public:
     void set(Pos pos_, Pos end_)
     {
         begin = pos = pos_;
+        start = begin;
         end = end_;
     }
 
@@ -44,6 +45,7 @@ public:
     bool get(Pos & token_begin, Pos & token_end)
     {
         /// Code from URLParser.
+
         if (pos == end)
             return false;
 
@@ -56,28 +58,23 @@ public:
                 ++pos;
 
             /** We will calculate the hierarchy only for URLs in which there is a protocol, and after it there are two slashes.
-             * (http, file - fit, mailto, magnet - do not fit), and after two slashes still at least something is there
-             * For the rest, simply return the full URL as the only element of the hierarchy.
+             * (http, file - fit, mailto, magnet - do not fit), and after two slashes still at least something is there.
+             * For the rest, just return an empty array.
              */
             if (pos == begin || pos == end || !(*pos++ == ':' && pos < end && *pos++ == '/' && pos < end && *pos++ == '/' && pos < end))
             {
                 pos = end;
-                token_begin = begin;
-                token_end = end;
-                return true;
+                return false;
             }
 
-            /// The domain for simplicity is everything that after the protocol and two slashes, until the next slash or `?` or `#`
+            /// The domain for simplicity is everything that after the protocol and the two slashes, until the next slash or `?` or `#`
             while (pos < end && !(*pos == '/' || *pos == '?' || *pos == '#'))
                 ++pos;
 
+            start = pos;
+
             if (pos != end)
                 ++pos;
-
-            token_begin = begin;
-            token_end = pos;
-
-            return true;
         }
 
         /// We go to the next `/` or `?` or `#`, skipping all those at the beginning.
@@ -91,7 +88,7 @@ public:
         if (pos != end)
             ++pos;
 
-        token_begin = begin;
+        token_begin = start;
         token_end = pos;
 
         return true;
@@ -99,12 +96,12 @@ public:
 };
 
 
-struct NameURLHierarchy { static constexpr auto name = "URLHierarchy"; };
-using FunctionURLHierarchy = FunctionTokens<URLHierarchyImpl>;
+struct NameURLPathHierarchy { static constexpr auto name = "URLPathHierarchy"; };
+using FunctionURLPathHierarchy = FunctionTokens<URLPathHierarchyImpl>;
 
-void registerFunctionURLHierarchy(FunctionFactory & factory)
+void registerFunctionURLPathHierarchy(FunctionFactory & factory)
 {
-    factory.registerFunction<FunctionURLHierarchy>();
+    factory.registerFunction<FunctionURLPathHierarchy>();
 }
 
 }
