+++ dbms/src/Functions/URL/config_functions_url.h.in
@@ -1,0 +1,5 @@
+#pragma once
+
+// .h autogenerated by cmake!
+
+#cmakedefine01 USE_GPERF
