+++ dbms/src/Functions/URL/decodeURLComponent.cpp
@@ -1,6 +1,5 @@
 #include <Common/hex.h>
 #include <Functions/FunctionFactory.h>
-#include <Functions/FunctionsURL.h>
 #include <Functions/FunctionStringToString.h>
 #include <common/find_symbols.h>
