+++ dbms/src/Functions/URL/tldLookup.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "config_functions.h"
+#include "config_functions_url.h"
 #if USE_GPERF
 // Definition of the class generated by gperf, present on gperf/tldLookup.gperf
 class tldLookupHash
