+++ dbms/src/Functions/config_functions.h.in
@@ -8,4 +8,3 @@
 #cmakedefine01 USE_HYPERSCAN
 #cmakedefine01 USE_SIMDJSON
 #cmakedefine01 USE_RAPIDJSON
-#cmakedefine01 USE_GPERF
